Show window initially by default
Given the new bin/catapult wrapper script we probably want for
consistency that the window is always shown when running the command,
regardless of whether an instance was already started prior or not. This
also probably removes the need to set Catapult to start when logging in
to the desktop, it'll just start and show when initially invoking the
keybinding set in the desktop/distro settings.