Add context managers to use a loading icon
This is a bit simpler than Gtk.Spinner, which doesn't seem to actually
spin if some code is blocking. Some kind of an async dance would
probably be required to make that work as intended.

There's a bizarre need for a tiny sleep here for the icon to actually
change. It could be that the 1/100 s that is enough on my system might
not be enough on someone else's system.

Closes #18